+++ .gitignore
@@ -1,4 +1,5 @@
-node_modules
-.parcel-cache
-dist
+.parcel-cache/
+node_modules/
+dist/
+lib/
 .DS_Store
