--- index.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Parcel sourcemap visualiser</title>
-  <link rel="stylesheet" href="./style.css">
-</head>
-<body>
-  <div id="app"></div>
-  
-  <script src="./index.tsx"></script>
-</body>
-</html>
