+++ website/index.tsx
@@ -4,6 +4,6 @@ import "regenerator-runtime/runtime";
 import React from "react";
 import ReactDOM from "react-dom";
 
-import App from "./components/app";
+import App from "./app";
 
 ReactDOM.render(<App />, document.getElementById("app"));
